+++ src/include/triangular_matrix_vector_kernel.hpp
@@ -12,9 +12,9 @@
 
 // fundamental real data type: 'float' or 'double'
 //using real_t = float;
-using real_t = float;
+using real_t = double;
 using vec_t = double;
-using mat_t = float;
+using mat_t = double;
 
 // number of bits to be used for exponent and mantissa
 #if defined(_BE)
